Fixed unsigned/signed mismatch in PacketCreator::addString
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@861 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/Common/PacketCreator.cpp
@@ -74,7 +74,7 @@ void PacketCreator::addString(const string &str, size_t len) {
 		std::cout << "ERROR: addString used with length shorter than string size." << std::endl; // TODO: Throw exception
 	}
 	strncpy_s((char *) packet + pos, MAX_LEN -pos, str.c_str(), slen);
-	for (int i = slen; i < len; i++) {
+	for (size_t i = slen; i < len; i++) {
 		packet[pos+i] = 0;
 	}
 	pos += len;
